pyproject.toml
[tool.poetry]
name = "prc"
version = "0.1.0"
description = ""
authors = ["Mazen Alotaibi <mazen.alotaibi@malaa.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.1"
clickhouse-sqlalchemy = {git = "https://github.com/malaa-tech/clickhouse-sqlalchemy", rev = "develop"}
sqlalchemy = "^2.0.25"
sqlalchemy-utils = "^0.41.1"
python-dotenv = "^1.0.0"
asynch = "^0.2.3"
pandas = "^2.1.4"
pytest = "^8.0.1"
pyarrow = "^15.0.0"
psycopg2-binary = "^2.9.9"
tqdm = "^4.66.2"
requests = "^2.31.0"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
